requirements.txt
streamlit
newspaper3k
transformers
requests
beautifulsoup4
